+++ Source/BuildPackages/Templates/Object_contains.pas
@@ -2,5 +2,5 @@ contains
   NovusObject in '..\..\Core\Object\NovusObject.pas',
   NovusBO in '..\..\Core\Object\NovusBO.pas',
   NovusBOField in '..\..\Core\Object\NovusBOField.pas',
-  NovusBOMap in '..\..\Core\NovusObject\NovusBOMap.pas',
-  NovusList in '..\..\Core\NovusObject\NovusList.pas';
+  NovusBOMap in '..\..\Core\Object\NovusBOMap.pas',
+  NovusList in '..\..\Core\Object\NovusList.pas';