+++ Source/Core/Console/NovusCommandLine.pas
@@ -3,7 +3,8 @@ unit NovusCommandLine;
 interface
 
 uses
-  NovusConsole, SysUtils, Classes, NovusList, System.StrUtils, System.RegularExpressions,
+  NovusConsole, SysUtils, Classes, NovusList, System.StrUtils,
+  System.RegularExpressions,
   System.Generics.Defaults;
 
 type
@@ -28,7 +29,7 @@ type
 
   end;
 
-   INovusCommandLineResultOptions =  interface
+  INovusCommandLineResultOptions = interface
     ['{394CA461-271D-46A2-89DD-9736F0909417}']
     function GetItems: tNovusList;
     procedure SetItems(Value: tNovusList);
@@ -36,11 +37,10 @@ type
     procedure Add(aItem: INovusCommandLineResultOption);
     function FirstOption: INovusCommandLineResultOption;
 
-    property Items: tNovuslist
-      read GetItems
-      write SetItems;
-  end;
+    function FindOptionByName(aOptionName: String): INovusCommandLineResultOption;
 
+    property Items: tNovusList read GetItems write SetItems;
+  end;
 
   INovusCommandLineResultCommand = interface
     ['{CC39CEF4-8289-4B74-9A9C-BCBCDAD0A730}']
@@ -48,17 +48,18 @@ type
     procedure SetCommandName(Value: string);
 
     function GetIsCommandOnly: boolean;
-    procedure SetIsCommandOnly(value: boolean);
+    procedure SetIsCommandOnly(Value: boolean);
     function GetOptions: INovusCommandLineResultOptions;
     procedure SetOptions(Value: INovusCommandLineResultOptions);
 
     property CommandName: string read GetCommandName write SetCommandName;
 
-    property IsCommandOnly: boolean read GetIsCommandOnly write SetIsCommandOnly;
+    property IsCommandOnly: boolean read GetIsCommandOnly
+      write SetIsCommandOnly;
 
-    property Options: INovusCommandLineResultOptions read GetOptions write SetOptions;
+    property Options: INovusCommandLineResultOptions read GetOptions
+      write SetOptions;
   end;
-
 
   INovusCommandLineResultCommands = interface
     ['{AFE16358-BAC6-47E1-95F1-5A8C9DBAB002}']
@@ -70,31 +71,31 @@ type
     function FirstCommand: INovusCommandLineResultCommand;
     function NextCommand: INovusCommandLineResultCommand;
     function GetIndex: Integer;
-    procedure SetIndex(value: integer);
+    procedure SetIndex(Value: Integer);
+    function FindCommandByName(aCommandName: string): INovusCommandLineResultCommand;
 
     function Exists(aCommandName: String): boolean;
 
     function GetCommands(aCommands: string): INovusCommandLineResultCommands;
 
-    property Items: tNovuslist
-      read GetItems
-      write SetItems;
+    property Items: tNovusList read GetItems write SetItems;
 
-    property Index: Integer
-      read GetIndex
-      write SetIndex;
+    property Index: Integer read GetIndex write SetIndex;
   end;
 
   INovusCommandLineResult = interface
     ['{4E2886DD-4120-4376-B778-0D803E621850}']
-    function GetErrors: Boolean;
-    procedure SetErrors(Value: Boolean);
+    function GetErrors: boolean;
+    procedure SetErrors(Value: boolean);
     function GetExitCode: Integer;
     procedure SetExitCode(Value: Integer);
     function GetErrorMessages: TStringlist;
     procedure SetErrorMessages(Value: TStringlist);
-    function GetIsCommandEmpty: Boolean;
-    procedure SetIsCommandEmpty(Value: Boolean);
+    function GetIsCommandEmpty: boolean;
+    procedure SetIsCommandEmpty(Value: boolean);
+    function GetIsHelpCommand: boolean;
+    procedure SetIsHelpCommand(Value: boolean);
+    procedure SetHelp(Value: String);
     function GetHelp: String;
     function GetCommands: INovusCommandLineResultCommands;
     procedure SetCommands(Value: INovusCommandLineResultCommands);
@@ -102,23 +103,31 @@ type
     procedure AddError(aErrorMessage: string; aExitCode: Integer = 0); overload;
     procedure AddError(aCommand: INovusCommandLineCommand); overload;
 
-    function FindFirstCommandwithOption(aCommandName: String): INovusCommandLineResultOption;
-    function FindFirstCommand(aCommandName: String): INovusCommandLineResultCommand;
+    function FindFirstCommandwithOption(aCommandName: String)
+      : INovusCommandLineResultOption;
+    function FindFirstCommand(aCommandName: String)
+      : INovusCommandLineResultCommand;
 
-    property IsCommandEmpty: Boolean read GetIsCommandEmpty
+    property IsCommandEmpty: boolean read GetIsCommandEmpty
       write SetIsCommandEmpty;
-    property Help: String read GetHelp;
 
-    property Errors: Boolean read GetErrors write SetErrors;
+    property IsHelpCommand: boolean read GetIsHelpCommand
+      write SetIsHelpCommand;
+
+    property Help: String read GetHelp write SetHelp;
+
+    property Errors: boolean read GetErrors write SetErrors;
     property ErrorMessages: TStringlist read GetErrorMessages
       write SetErrorMessages;
 
     property ExitCode: Integer read GetExitCode write SetExitCode;
 
-    property Commands: INovusCommandLineResultCommands read GetCommands write SetCommands;
+    property Commands: INovusCommandLineResultCommands read GetCommands
+      write SetCommands;
   end;
 
-  TNovusCommandLineResultOption = class(TSingletonImplementation, INovusCommandLineResultOption)
+  TNovusCommandLineResultOption = class(TSingletonImplementation,
+    INovusCommandLineResultOption)
   private
     fsOptionName: String;
     fsValue: String;
@@ -132,20 +141,19 @@ type
     property Value: string read GetValue write SetValue;
   end;
 
-
-
-  TNovusCommandLineResultCommand = class(TSingletonImplementation, INovusCommandLineResultCommand)
+  TNovusCommandLineResultCommand = class(TSingletonImplementation,
+    INovusCommandLineResultCommand)
   protected
     fOptions: INovusCommandLineResultOptions;
     fsCommandName: string;
-    fbIsCommandOnly: Boolean;
+    fbIsCommandOnly: boolean;
   private
     fbIsCommand: boolean;
     function GetCommandName: string;
     procedure SetCommandName(Value: string);
 
     function GetIsCommandOnly: boolean;
-    procedure SetIsCommandOnly(value: boolean);
+    procedure SetIsCommandOnly(Value: boolean);
 
     function GetOptions: INovusCommandLineResultOptions;
     procedure SetOptions(Value: INovusCommandLineResultOptions);
@@ -155,16 +163,17 @@ type
 
     property CommandName: string read GetCommandName write SetCommandName;
 
+    property IsCommandOnly: boolean read GetIsCommandOnly
+      write SetIsCommandOnly;
 
-    property IsCommandOnly: Boolean read GetIsCommandOnly write SetIsCommandOnly;
-
-
-    property Options: INovusCommandLineResultOptions read GetOptions write SetOptions;
+    property Options: INovusCommandLineResultOptions read GetOptions
+      write SetOptions;
   end;
 
-  TNovusCommandLineResultOptions = class(TSingletonImplementation, INovusCommandLineResultOptions)
+  TNovusCommandLineResultOptions = class(TSingletonImplementation,
+    INovusCommandLineResultOptions)
   protected
-    FNovusCommandLineResultOptions: TNovusList;
+    FNovusCommandLineResultOptions: tNovusList;
   private
     function GetItems: tNovusList;
     procedure SetItems(Value: tNovusList);
@@ -174,21 +183,21 @@ type
 
     procedure Add(aItem: INovusCommandLineResultOption);
     function FirstOption: INovusCommandLineResultOption;
+    function FindOptionByName(aOptionName: String): INovusCommandLineResultOption;
 
-    property Items: tNovuslist
-      read GetItems
-      write SetItems;
+    property Items: tNovusList read GetItems write SetItems;
   end;
 
-  TNovusCommandLineResultCommands = class(TSingletonImplementation, INovusCommandLineResultCommands)
+  TNovusCommandLineResultCommands = class(TSingletonImplementation,
+    INovusCommandLineResultCommands)
   protected
-    FNovusCommandLineResultCommands: TNovusList;
+    FNovusCommandLineResultCommands: tNovusList;
   private
     fiIndex: Integer;
     function GetItems: tNovusList;
     procedure SetItems(Value: tNovusList);
     function GetIndex: Integer;
-    procedure SetIndex(value: integer);
+    procedure SetIndex(Value: Integer);
   public
     constructor Create;
     destructor Destroy;
@@ -196,38 +205,42 @@ type
     function GetCommands(aCommandName: string): INovusCommandLineResultCommands;
     function FirstCommand: INovusCommandLineResultCommand;
     function NextCommand: INovusCommandLineResultCommand;
+    function FindCommandByName(aCommandName: string): INovusCommandLineResultCommand;
 
     procedure Add(aItem: INovusCommandLineResultCommand);
 
     function Exists(aCommandName: String): boolean;
 
-    property Items: tNovuslist read GetItems write SetItems;
+    property Items: tNovusList read GetItems write SetItems;
 
-    property Index: Integer
-      read GetIndex
-      write SetIndex;
+    property Index: Integer read GetIndex write SetIndex;
   end;
 
-  TNovusCommandLineResult = class(TSingletonImplementation, INovusCommandLineResult)
+  TNovusCommandLineResult = class(TSingletonImplementation,
+    INovusCommandLineResult)
   protected
     fCommands: INovusCommandLineResultCommands;
     fiExitCode: Integer;
-    fbIsCommandEmpty: Boolean;
-    fbErrors: Boolean;
+    fbIsCommandEmpty: boolean;
+    fbIsHelpCommand: boolean;
+    fbErrors: boolean;
     fsHelp: string;
     fErrorMessages: TStringlist;
   private
-    function GetErrors: Boolean;
-    procedure SetErrors(Value: Boolean);
+    function GetErrors: boolean;
+    procedure SetErrors(Value: boolean);
+    procedure SetHelp(Value: String);
     function GetHelp: String;
     function GetExitCode: Integer;
     procedure SetExitCode(Value: Integer);
     function GetErrorMessages: TStringlist;
     procedure SetErrorMessages(Value: TStringlist);
-    function GetIsCommandEmpty: Boolean;
-    procedure SetIsCommandEmpty(Value: Boolean);
+    function GetIsCommandEmpty: boolean;
+    procedure SetIsCommandEmpty(Value: boolean);
     function GetCommands: INovusCommandLineResultCommands;
     procedure SetCommands(Value: INovusCommandLineResultCommands);
+    function GetIsHelpCommand: boolean;
+    procedure SetIsHelpCommand(Value: boolean);
   public
     constructor Create;
     destructor Destroy;
@@ -235,30 +248,34 @@ type
     procedure AddError(aErrorMessage: string; aExitCode: Integer = 0); overload;
     procedure AddError(aCommand: INovusCommandLineCommand); overload;
 
-    function FindFirstCommand(aCommandName: String): INovusCommandLineResultCommand;
-    function FindFirstCommandwithOption(aCommandName: String): INovusCommandLineResultOption;
+    function FindFirstCommand(aCommandName: String)
+      : INovusCommandLineResultCommand;
+    function FindFirstCommandwithOption(aCommandName: String)
+      : INovusCommandLineResultOption;
 
-    property IsCommandEmpty: Boolean read GetIsCommandEmpty
+    property IsCommandEmpty: boolean read GetIsCommandEmpty
       write SetIsCommandEmpty;
 
-    property Help: String read GetHelp;
-    property Errors: Boolean read GetErrors write SetErrors;
+    property IsHelpCommand: boolean read GetIsHelpCommand
+      write SetIsHelpCommand;
+
+    property Help: String read GetHelp write SetHelp;
+    property Errors: boolean read GetErrors write SetErrors;
     property ErrorMessages: TStringlist read GetErrorMessages
       write SetErrorMessages;
 
-    property Commands: INovusCommandLineResultCommands read GetCommands write SetCommands;
+    property Commands: INovusCommandLineResultCommands read GetCommands
+      write SetCommands;
 
-    property ExitCode: Integer
-       read fiExitCode
-       write fiExitCode;
+    property ExitCode: Integer read fiExitCode write fiExitCode;
   end;
 
   INovusCommandLineOption = interface
     ['{2195EF48-095A-4AF1-94F6-3B8B953584CF}']
     function GetOptionName: string;
     procedure SetOptionName(Value: string);
-    procedure SetRequired(Value: Boolean);
-    function GetRequired: Boolean;
+    procedure SetRequired(Value: boolean);
+    function GetRequired: boolean;
     function GetHelp: string;
     procedure SetHelp(Value: string);
     function GetValue: string;
@@ -269,32 +286,32 @@ type
     function GetExitCode: Integer;
     procedure SetExitCode(Value: Integer);
 
-    function Execute: Boolean;
+    function Execute: boolean;
     procedure AddError(aErrorMessage: string);
 
     property OptionName: string read GetOptionName write SetOptionName;
-    property Required: Boolean read GetRequired write SetRequired;
+    property Required: boolean read GetRequired write SetRequired;
     property Help: string read GetHelp write SetHelp;
     property Value: string read GetValue write SetValue;
     property ErrorMessages: TStringlist read GetErrorMessages
       write SetErrorMessages;
-    property ExitCode: Integer
-       read GetExitCode write SetExitCode;
+    property ExitCode: Integer read GetExitCode write SetExitCode;
   end;
 
-  tNovusCommandLineOption = class(TSingletonImplementation, INovusCommandLineOption)
+  tNovusCommandLineOption = class(TSingletonImplementation,
+    INovusCommandLineOption)
   protected
     fiExitCode: Integer;
     fsOptionName: string;
     fsHelp: string;
-    fbRequired: Boolean;
+    fbRequired: boolean;
     fsValue: string;
     fErrorMessages: TStringlist;
   private
     function GetOptionName: string;
     procedure SetOptionName(Value: string);
-    procedure SetRequired(Value: Boolean);
-    function GetRequired: Boolean;
+    procedure SetRequired(Value: boolean);
+    function GetRequired: boolean;
     function GetHelp: string;
     procedure SetHelp(Value: string);
     function GetValue: string;
@@ -309,17 +326,16 @@ type
     constructor Create; virtual;
     destructor Destroy;
 
-    function Execute: Boolean; virtual;
+    function Execute: boolean; virtual;
     procedure AddError(aErrorMessage: string);
 
     property OptionName: string read GetOptionName write SetOptionName;
-    property Required: Boolean read GetRequired write SetRequired;
+    property Required: boolean read GetRequired write SetRequired;
     property Help: string read GetHelp write SetHelp;
     property Value: string read GetValue write SetValue;
     property ErrorMessages: TStringlist read GetErrorMessages
       write SetErrorMessages;
-    property ExitCode : integer
-       read GetExitCode write SetExitCode;
+    property ExitCode: Integer read GetExitCode write SetExitCode;
   end;
 
   INovusCommandLineCommand = interface
@@ -330,25 +346,30 @@ type
     procedure SetShortCommandName(Value: string);
     function GetHelp: string;
     procedure SetHelp(Value: string);
-    procedure SetRequired(Value: Boolean);
-    function GetRequired: Boolean;
+    procedure SetRequired(Value: boolean);
+    function GetRequired: boolean;
+
+    procedure SetHelpCommand(Value: boolean);
+    function GetHelpCommand: boolean;
+
     function GetErrorMessages: TStringlist;
     procedure SetErrorMessages(Value: TStringlist);
-    function GetOptionList: tNovuslist;
-    procedure SetOptionList(Value: tNovuslist);
+    function GetOptionsList: tNovusList;
+    procedure SetOptionsList(Value: tNovusList);
     function GetExitCode: Integer;
     procedure SetExitCode(Value: Integer);
-    function GetCommandList: tNovuslist;
-    procedure SetCommandList(Value: tNovuslist);
+    function GetCommandList: tNovusList;
+    procedure SetCommandList(Value: tNovusList);
 
-    function IsOptionsRequried: Boolean;
+    function GetOptionsCount: Integer;
 
     property CommandName: string read GetCommandName write SetCommandName;
     property ShortCommandName: string read GetShortCommandName
       write SetShortCommandName;
     property Help: string read GetHelp write SetHelp;
 
-    property Required: Boolean read GetRequired write SetRequired;
+    property Required: boolean read GetRequired write SetRequired;
+    property HelpCommand: boolean read GetHelpCommand write SetHelpCommand;
 
     procedure RegisterObject(aObject: TObject);
     function FindRegisterObject(aClassName: String): TObject;
@@ -356,15 +377,16 @@ type
     function FindCommandName(aCommandName: string): INovusCommandLineCommand;
 
     function FindOptionByName(aOptionName: string): INovusCommandLineOption;
-    function IsOptionsExists: Boolean;
+
     procedure AddError(aErrorMessage: string; aExitCode: Integer = 0);
 
-    function Execute: Boolean;
-    function Parse: Boolean; overload;
+    function Execute: boolean;
+    function Parse: boolean;
 
+    function IsOptionsRequired: boolean;
 
     function RegisterOption(const aOptionName: string; const aHelp: String;
-      const aRequired: Boolean; aCommandLineOption: tNovusCommandLineOption)
+      const aRequired: boolean; aCommandLineOption: tNovusCommandLineOption)
       : tNovusCommandLineOption;
 
     property ErrorMessages: TStringlist read GetErrorMessages
@@ -372,40 +394,47 @@ type
 
     property ExitCode: Integer read GetExitCode write SetExitCode;
 
-    property OptionList: tNovuslist read GetOptionList write SetOptionList;
+    property OptionsList: tNovusList read GetOptionsList write SetOptionsList;
 
-    property CommandList: tNovuslist read GetCommandList write SetCommandList;
+    property CommandList: tNovusList read GetCommandList write SetCommandList;
+
+    property OptionsCount: Integer read GetOptionsCount;
 
   end;
 
-  tNovusCommandLineCommand = class(TSingletonImplementation, INovusCommandLineCommand)
+  tNovusCommandLineCommand = class(TSingletonImplementation,
+    INovusCommandLineCommand)
   protected
     fiExitCode: Integer;
     fErrorMessages: TStringlist;
-    fbRequired: Boolean;
+    fbRequired: boolean;
+    fbHelpCommand: boolean;
     fsHelp: string;
     fsCommandName: String;
     fsShortCommandName: String;
-    fOptionList: tNovuslist;
-    fCommandList: tNovuslist;
-    fObjectList: tNovuslist;
+    fOptionsList: tNovusList;
+    fCommandList: tNovusList;
+    fObjectList: tNovusList;
   private
-    function GetOptionList: tNovuslist;
-    procedure SetOptionList(Value: tNovuslist);
+    function GetOptionsList: tNovusList;
+    procedure SetOptionsList(Value: tNovusList);
     function GetCommandName: string;
     procedure SetCommandName(Value: string);
     function GetShortCommandName: string;
     procedure SetShortCommandName(Value: string);
     function GetHelp: string;
     procedure SetHelp(Value: string);
-    procedure SetRequired(Value: Boolean);
-    function GetRequired: Boolean;
+    procedure SetRequired(Value: boolean);
+    function GetRequired: boolean;
+    procedure SetHelpCommand(Value: boolean);
+    function GetHelpCommand: boolean;
     function GetErrorMessages: TStringlist;
     procedure SetErrorMessages(Value: TStringlist);
     function GetExitCode: Integer;
     procedure SetExitCode(Value: Integer);
-    function GetCommandList: tNovuslist;
-    procedure SetCommandList(Value: tNovuslist);
+    function GetCommandList: tNovusList;
+    procedure SetCommandList(Value: tNovusList);
+    function GetOptionsCount: Integer;
   public
     constructor Create; virtual;
     destructor Destroy;
@@ -416,15 +445,14 @@ type
     property ShortCommandName: string read GetShortCommandName
       write SetShortCommandName;
     property Help: string read GetHelp write SetHelp;
-    property Required: Boolean read GetRequired write SetRequired;
+    property Required: boolean read GetRequired write SetRequired;
+    property HelpCommand: boolean read GetHelpCommand write SetHelpCommand;
 
-    function Execute: Boolean; virtual;
-    function Parse: Boolean;
-  
+    function Execute: boolean; virtual;
+    function Parse: boolean;
+
     function FindOptionByName(aOptionName: string): INovusCommandLineOption;
-    function IsOptionsExists: Boolean;
-
-    function IsOptionsRequried: Boolean;
+    function IsOptionsRequired: boolean;
 
     procedure AddError(aErrorMessage: string; aExitCode: Integer = 0);
 
@@ -432,17 +460,19 @@ type
     function FindRegisterObject(aClassName: String): TObject;
 
     function RegisterOption(const aOptionName: string; const aHelp: String;
-      const aRequired: Boolean; aCommandLineOption: tNovusCommandLineOption)
+      const aRequired: boolean; aCommandLineOption: tNovusCommandLineOption)
       : tNovusCommandLineOption;
 
-    property OptionList: tNovuslist read GetOptionList write SetOptionList;
+    property OptionsList: tNovusList read GetOptionsList write SetOptionsList;
 
     property ErrorMessages: TStringlist read GetErrorMessages
       write SetErrorMessages;
 
     property ExitCode: Integer read GetExitCode write SetExitCode;
 
-    property CommandList: tNovuslist read GetCommandList write SetCommandList;
+    property CommandList: tNovusList read GetCommandList write SetCommandList;
+
+    property OptionsCount: Integer read GetOptionsCount;
 
   end;
 
@@ -454,12 +484,13 @@ type
   tNovusCommandLine = class(TSingletonImplementation, INovusCommandLine)
   protected
   class var
-    fCommandList: tNovuslist;
+    fCommandList: tNovusList;
     FParamStrList: TStringlist;
     fiParamIndex: Integer;
   private
     class function ParamStrToStringList: TStringlist;
-    class function InternalParse(aOptions: tNovusCommandOptions): TNovusCommandLineResult;
+    class function InternalParse(aOptions: tNovusCommandOptions)
+      : TNovusCommandLineResult;
     class function FindCommandName(aCommandName: string)
       : INovusCommandLineCommand;
   public
@@ -467,21 +498,24 @@ type
     class destructor Destroy;
 
     class function Execute: TNovusCommandLineResult; overload;
-    class function Execute(aOptions: tNovusCommandOptions): TNovusCommandLineResult; overload;
+    class function Execute(aOptions: tNovusCommandOptions)
+      : TNovusCommandLineResult; overload;
 
     class function RegisterCommand(const aCommandName: string;
       const aShortCommandName: string; const aHelp: String;
-      const aRequired: Boolean; aCommandLineCommand: tNovusCommandLineCommand)
+      const aRequired: boolean; aCommandLineCommand: tNovusCommandLineCommand)
       : tNovusCommandLineCommand;
 
-
+    class function RegisterHelpCommand(const aCommandName: string;
+      const aShortCommandName: string; const aHelp: String;
+      aCommandLineCommand: tNovusCommandLineCommand): tNovusCommandLineCommand;
   end;
 
 implementation
 
 class constructor tNovusCommandLine.Create;
 begin
-  fCommandList := tNovuslist.Create;
+  fCommandList := tNovusList.Create;
 end;
 
 class destructor tNovusCommandLine.Destroy;
@@ -494,21 +528,21 @@ end;
 
 class function tNovusCommandLine.Execute: TNovusCommandLineResult;
 Var
-  FOptions: tNovusCommandOptions;
+  fOptions: tNovusCommandOptions;
 begin
   Result := NIL;
 
-  FOptions.CommandEmptyExitCode := -1;
-  FOptions.CommandLineExitCode := -2;
-  FOptions.CommandLineOptionExitCode := -3;
-
+  fOptions.CommandEmptyExitCode := -1;
+  fOptions.CommandLineExitCode := -2;
+  fOptions.CommandLineOptionExitCode := -3;
 
   FParamStrList := tNovusCommandLine.ParamStrToStringList;
 
-  Result := tNovusCommandLine.InternalParse(FOptions);
+  Result := tNovusCommandLine.InternalParse(fOptions);
 end;
 
-class function tNovusCommandLine.Execute(aOptions: tNovusCommandOptions): TNovusCommandLineResult;
+class function tNovusCommandLine.Execute(aOptions: tNovusCommandOptions)
+  : TNovusCommandLineResult;
 begin
   Result := NIL;
 
@@ -516,8 +550,6 @@ begin
 
   Result := tNovusCommandLine.InternalParse(aOptions);
 end;
-
-
 
 class function tNovusCommandLine.ParamStrToStringList: TStringlist;
 Var
@@ -538,55 +570,54 @@ Var
   fCommand: tNovusCommandLineCommand;
   i: Integer;
 
-
   function IsAplhaNumbicString(aString: String): boolean;
   begin
-    result := false;
+    Result := false;
 
-    Result := Not System.RegularExpressions.TRegEx.IsMatch(aString, '[^A-Za-z0-9\.]+');
+    Result := Not System.RegularExpressions.TRegEx.IsMatch(aString,
+      '[^A-Za-z0-9\.]+');
   end;
-
-
 
 begin
   Result := NIL;
 
-  if not  IsAplhaNumbicString(aCommandName) then Exit;
+  if not IsAplhaNumbicString(aCommandName) then
+    Exit;
 
   For i := 0 to fCommandList.count - 1 do
   begin
-    if fCommandList.items[i] Is tNovusCommandLineCommand then
+    if fCommandList.Items[i] Is tNovusCommandLineCommand then
+    begin
+      fCommand := fCommandList.Items[i] as tNovusCommandLineCommand;
+
+      if (Uppercase(fCommand.CommandName) = Uppercase(aCommandName)) or
+        (Uppercase(fCommand.ShortCommandName) = Uppercase(aCommandName)) then
       begin
-        fCommand := fCommandList.items[i] as tNovusCommandLineCommand;
+        Result := fCommand;
 
-        if (Uppercase(fCommand.CommandName) = Uppercase(aCommandName)) or
-          (Uppercase(fCommand.ShortCommandName) = Uppercase(aCommandName)) then
-        begin
-          Result := fCommand;
-
-          break;
-        end;
-     end;
+        break;
+      end;
+    end;
 
   end;
 end;
 
-
-
-
-class function tNovusCommandLine.InternalParse(aOptions: tNovusCommandOptions): TNovusCommandLineResult;
+class function tNovusCommandLine.InternalParse(aOptions: tNovusCommandOptions)
+  : TNovusCommandLineResult;
 Var
+  lbIsHelpCommand: boolean;
   lsParamValue: string;
   lCommand, lLastCommand: INovusCommandLineCommand;
   liOptionIndex: Integer;
   lOption: tNovusCommandLineOption;
-  fbOptionsRequried: Boolean;
-  lbIsCommand: Boolean;
+  fbOptionsExists: boolean;
+  lbIsCommand: boolean;
   lsLastParamValue: String;
   fResultCommand: INovusCommandLineResultCommand;
   fResultOption: INovusCommandLineResultOption;
-  I: integer;
+  i: Integer;
   fCommand: tNovusCommandLineCommand;
+  fTempResultOptions: TNovusCommandLineResultOptions;
 begin
   Result := TNovusCommandLineResult.Create;
 
@@ -595,17 +626,20 @@ begin
     Result.IsCommandEmpty := true;
 
     Result.Errors := true;
-  
+
     Result.ExitCode := aOptions.CommandEmptyExitCode;
 
     Exit;
   end;
 
+  fTempResultOptions := TNovusCommandLineResultOptions.Create;
+
   fiParamIndex := 0;
   liOptionIndex := 0;
-  fbOptionsRequried := false;
+  fbOptionsExists := false;
   lLastCommand := NIL;
-  While (fiParamIndex <= FParamStrList.count - 1) do
+  lbIsHelpCommand := false;
+  While (fiParamIndex < FParamStrList.count) do
   begin
     lsParamValue := Trim(FParamStrList.Strings[fiParamIndex]);
 
@@ -629,72 +663,92 @@ begin
     if Trim(lsLastParamValue) <> Trim(lsParamValue) then
       lbIsCommand := true;
 
-
     lCommand := tNovusCommandLine.FindCommandName(lsParamValue);
-    if Assigned(lCommand) and (lLastCommand <> lCommand) then
+    if Assigned(lCommand) and (lLastCommand <> lCommand) and
+      (lbIsHelpCommand = false) then
     begin
       lLastCommand := lCommand;
       liOptionIndex := 0;
-      fbOptionsRequried := lLastCommand.IsOptionsRequried;
+      fbOptionsExists := (lLastCommand.OptionsCount <> 0);
+
+      if lLastCommand.HelpCommand then
+        lbIsHelpCommand := true;
     end
-    else
-    if Assigned(lLastCommand) then
+    else if Assigned(lLastCommand) then
+    begin
+      // fbOptionsRequried := false;
+      lOption := nil;
+      if lLastCommand.OptionsCount > 0 then
       begin
-        // fbOptionsRequried := false;
-        lOption := nil;
-        if lLastCommand.IsOptionsExists then
+        if (liOptionIndex <= lLastCommand.OptionsList.count - 1) then
         begin
-          if (liOptionIndex <= lLastCommand.OptionList.count - 1) then
+          lOption := tNovusCommandLineOption(lLastCommand.OptionsList.Items
+            [liOptionIndex]);
+
+          lOption.Value := lsParamValue;
+
+          if Trim(Result.Help) = '' then
+            Result.Help := lLastCommand.Help;
+
+          if (lOption.Value = '') and (lOption.Required = true) then
           begin
-            lOption := tNovusCommandLineOption
-              (lLastCommand.OptionList.items[liOptionIndex]);
+             Result.AddError('Required Option [' + lOption.OptionName +
+              '] was not specified');
 
-            lOption.Value := lsParamValue;
+            if Result.ExitCode = 0 then
+              Result.ExitCode := lOption.ExitCode;
 
-            if (lOption.Value = '') and (lOption.Required = true) then
-            begin
-              Result.AddError('Required Option [' + lOption.OptionName +
-                '] was not specified');
+            break;
+          end
+          else if not lOption.Execute then
+          begin
+            Result.AddError(lOption.ErrorMessages.Text);
 
-              result.ExitCode := aOptions.CommandLineOptionExitCode;
+            if Result.ExitCode = 0 then
+              Result.ExitCode := lOption.ExitCode;
 
-              break;
-            end
-            else if not lOption.Execute then
-            begin
-              Result.AddError(lOption.ErrorMessages.Text);
-
-              result.ExitCode := lOption.ExitCode;
-              if result.ExitCode = 0 then
-                result.ExitCode := aOptions.CommandLineOptionExitCode;
-
-              break;
-            end;
+            break;
           end
           else
           begin
-            Result.AddError('Unknown command line option: ' +
-              FParamStrList.Strings[fiParamIndex]);
+            fResultOption := TNovusCommandLineResultOption.Create;
 
-            result.ExitCode := aOptions.CommandLineOptionExitCode;
+            fResultOption.OptionName := lOption.OptionName;
+            fResultOption.Value := lOption.Value;
 
-            break;
+            fTempResultOptions.Add(fResultOption);
           end;
-
-          inc(liOptionIndex);
         end
         else
         begin
           Result.AddError('Unknown command line option: ' +
             FParamStrList.Strings[fiParamIndex]);
 
-          result.ExitCode := aOptions.CommandLineOptionExitCode;
+          if Result.ExitCode = 0 then
+            Result.ExitCode := lLastCommand.ExitCode;
 
           break;
         end;
 
+        inc(liOptionIndex);
+      end
+      else
+      begin
+        Result.AddError('Unknown command line option: ' + FParamStrList.Strings
+          [fiParamIndex]);
 
-      if not Result.Errors then
+        if Result.ExitCode = 0 then
+          Result.ExitCode := aOptions.CommandLineOptionExitCode;
+
+        break;
+      end;
+
+      if (liOptionIndex > lLastCommand.OptionsList.count - 1) or
+        (fbOptionsExists = false) then
+      begin
+        liOptionIndex := 0;
+
+        if not Result.Errors then
         begin
           if not lLastCommand.Execute then
           begin
@@ -702,93 +756,196 @@ begin
 
             Result.ExitCode := lLastCommand.ExitCode;
             if Result.ExitCode = 0 then
-               Result.ExitCode := aOptions.CommandLineExitCode;
+              Result.ExitCode := aOptions.CommandLineExitCode;
+
+            if Trim(lLastCommand.Help) <> '' then
+              Result.Help := lLastCommand.Help;
 
             break;
           end
-        else
+          else
           begin
             fResultCommand := TNovusCommandLineResultCommand.Create;
 
             fResultCommand.CommandName := lLastCommand.CommandName;
 
-            if Assigned(lOption) then
+            if fTempResultOptions.Items.count > 0 then
+            begin
+              fResultCommand.Options.Items.CopyFrom(fTempResultOptions.Items);
+
+              fTempResultOptions.Items.Clear;
+            end
+            else If fResultCommand.Options.Items.count = 0 then
+              fResultCommand.IsCommandOnly := true;
+
+            (*
+              if Assigned(lOption) then
               begin
-                fResultOption:= TNovusCommandLineResultOption.Create;
+              fResultOption:= TNovusCommandLineResultOption.Create;
 
-                fResultOption.OptionName := lOption.OptionName;
-                fResultOption.Value := lOption.Value;
+              fResultOption.OptionName := lOption.OptionName;
+              fResultOption.Value := lOption.Value;
 
-                fResultCommand.Options.Add(fResultOption);
+              fResultCommand.Options.Add(fResultOption);
               end
-            else fResultCommand.IsCommandOnly := true;
+              else fResultCommand.IsCommandOnly := true;
+            *)
+
+            Result.ExitCode := lLastCommand.ExitCode;
+
+            if Trim(lLastCommand.Help) <> '' then
+              Result.Help := lLastCommand.Help;
 
             Result.Commands.Add(fResultCommand);
 
             lLastCommand := NIL;
           end;
+        end;
 
-      end
+      end;
+      (*
       else
       begin
         Result.AddError('Unknown command line option: ' + FParamStrList.Strings
           [fiParamIndex]);
 
-       Result.ExitCode := aOPtions.CommandLineOptionExitCode;
+        Result.ExitCode := aOptions.CommandLineOptionExitCode;
 
         break;
       end;
+      *)
     end
-    else
-      if lbIsCommand then
-        begin
-          Result.AddError('Unknown command line: ' + lsLastParamValue);
+    else if lbIsCommand then
+    begin
+      Result.AddError('Unknown command line: ' + lsLastParamValue);
 
-          Result.ExitCode := aOPtions.CommandLineExitCode;
+      Result.ExitCode := aOptions.CommandLineExitCode;
 
-          break;
-        end;
+      if Assigned(lLastCommand) then
+       if Trim(Result.Help) = '' then
+          Result.Help := lLastCommand.Help;
 
-    if not fbOptionsRequried and lbIsCommand then
-      begin
-        fResultCommand := TNovusCommandLineResultCommand.Create;
+      break;
+    end;
 
-        fResultCommand.CommandName := lLastCommand.CommandName;
+    if not fbOptionsExists and lbIsCommand then
+    begin
+      fResultCommand := TNovusCommandLineResultCommand.Create;
 
-        fResultCommand.IsCommandOnly := true;
+      fResultCommand.CommandName := lLastCommand.CommandName;
 
-        Result.Commands.Add(fResultCommand);
+      fResultCommand.IsCommandOnly := true;
 
-        lLastCommand := NIL;
-      end;
+      Result.Commands.Add(fResultCommand);
 
+      lLastCommand := NIL;
+    end;
 
     inc(fiParamIndex);
   end;
 
-
-  // Check Required Command
-  For i := 0 to fCommandList.count - 1 do
-  begin
-    fCommand := fCommandList.items[i] as tNovusCommandLineCommand;
-
-    if fCommand.Required then
+  If (fiParamIndex = FParamStrList.count) then
+    if Assigned(lLastCommand) then
+    begin
+      if fbOptionsExists then
       begin
-        if not Result.Commands.Exists(fCommand.CommandName) then
+        if lLastCommand.OptionsCount > 0 then
+        begin
+          liOptionIndex := 0;
+
+          fResultCommand := NIl;
+
+          if lLastCommand.IsOptionsRequired then
           begin
-            Result.AddError('Required Command [' + fCommand.CommandName +'] was not specified');
-            Result.ExitCode := aOPtions.CommandLineExitCode;
+            for I := 0 to lLastCommand.OptionsList.Count -1 do
+              begin
+                lOption := tNovusCommandLineOption(lLastCommand.OptionsList.Items
+                      [I]);
+
+                if lOption.Required then
+                  begin
+                    fResultCommand :=  Result.Commands.FindCommandByName(lLastCommand.CommandName);
+
+                    if Assigned(fResultCommand) then
+                      begin
+                        if Not Assigned(fResultCommand.Options.FindOptionByName(lOption.OptionName )) then
+                          begin
+                            Result.AddError('Required Option [' + lOption.OptionName +
+                                 '] was not specified');
+
+                          end;
+                      end
+                    else
+                       begin
+                         if FTempResultOptions.Items.Count > 0 then
+                           begin
+                             if Not Assigned(FTempResultOptions.FindOptionByName(lOption.OptionName )) then
+                                begin
+                                  Result.AddError('Required Option [' + lOption.OptionName +
+                                       '] was not specified');
+
+                                end;
+                           end
+                         else
+                           begin
+                             Result.AddError('Required Option [' + lOption.OptionName +
+                                 '] was not specified');
+
+
+                           end;
+                       end;
+                  end;
+              end;
           end;
 
+          Result.ExitCode := aOptions.CommandLineOptionExitCode;
+
+          if Trim(lLastCommand.Help) <> '' then
+            Result.Help := lLastCommand.Help;
+        end;
       end;
-  end;
+    end;
+
+  // Check Required Command
+  if lbIsHelpCommand = false then
+  begin
+    For i := 0 to fCommandList.count - 1 do
+    begin
+      fCommand := fCommandList.Items[i] as tNovusCommandLineCommand;
+
+      if Assigned(lLastCommand) then
+        if (lLastCommand.CommandName = fCommand.CommandName) then
+          break;
+
+      if fCommand.Required then
+      begin
+        if not Result.Commands.Exists(fCommand.CommandName) then
+        begin
+          Result.AddError('Required Command [' + fCommand.CommandName +
+            '] was not specified');
+          Result.ExitCode := aOptions.CommandLineExitCode;
+
+          if assigned(lLastCommand) then
+            if Trim(lLastCommand.Help) <> '' then
+               Result.Help := lLastCommand.Help;
+
+          break;
+        end;
+
+      end;
+    end;
+  end
+  else
+    Result.IsHelpCommand := lbIsHelpCommand;
+
+  fTempResultOptions.Free;
+
 end;
 
 // tNovusCommandLine
-
 class function tNovusCommandLine.RegisterCommand(const aCommandName: string;
   const aShortCommandName: string; const aHelp: String;
-  const aRequired: Boolean; aCommandLineCommand: tNovusCommandLineCommand)
+  const aRequired: boolean; aCommandLineCommand: tNovusCommandLineCommand)
   : tNovusCommandLineCommand;
 begin
   if not Assigned(aCommandLineCommand) then
@@ -798,6 +955,25 @@ begin
   aCommandLineCommand.ShortCommandName := aShortCommandName;
   aCommandLineCommand.Help := aHelp;
   aCommandLineCommand.Required := aRequired;
+  aCommandLineCommand.HelpCommand := false;
+
+  aCommandLineCommand.CommandList := fCommandList;
+
+  fCommandList.Add(aCommandLineCommand);
+end;
+
+class function tNovusCommandLine.RegisterHelpCommand(const aCommandName: string;
+  const aShortCommandName: string; const aHelp: String;
+  aCommandLineCommand: tNovusCommandLineCommand): tNovusCommandLineCommand;
+begin
+  if not Assigned(aCommandLineCommand) then
+    aCommandLineCommand := tNovusCommandLineCommand.Create;
+
+  aCommandLineCommand.CommandName := aCommandName;
+  aCommandLineCommand.ShortCommandName := aShortCommandName;
+  aCommandLineCommand.Help := aHelp;
+  aCommandLineCommand.Required := false;
+  aCommandLineCommand.HelpCommand := true;
 
   aCommandLineCommand.CommandList := fCommandList;
 
@@ -833,14 +1009,14 @@ begin
   fsCommandName := Value;
 end;
 
-function tNovusCommandLineCommand.GetOptionList: tNovuslist;
+function tNovusCommandLineCommand.GetOptionsList: tNovusList;
 begin
-  Result := fOptionList;
+  Result := fOptionsList;
 end;
 
-procedure tNovusCommandLineCommand.SetOptionList(Value: tNovuslist);
+procedure tNovusCommandLineCommand.SetOptionsList(Value: tNovusList);
 begin
-  fOptionList := Value;
+  fOptionsList := Value;
 end;
 
 function tNovusCommandLineCommand.GetShortCommandName: string;
@@ -853,29 +1029,21 @@ begin
   fsShortCommandName := Value;
 end;
 
-function tNovusCommandLineCommand.Execute: Boolean;
+function tNovusCommandLineCommand.Execute: boolean;
 begin
   Result := true;
 end;
 
-function tNovusCommandLineCommand.IsOptionsExists: Boolean;
-begin
-  Result := (fOptionList.count <> 0);
-end;
-
-
-
-
-function tNovusCommandLineCommand.IsOptionsRequried: Boolean;
+function tNovusCommandLineCommand.IsOptionsRequired: boolean;
 var
   lOption: tNovusCommandLineOption;
   i: Integer;
 begin
   Result := false;
 
-  for i := 0 to OptionList.count - 1 do
+  for i := 0 to OptionsList.count - 1 do
   begin
-    lOption := tNovusCommandLineOption(OptionList.items[i]);
+    lOption := tNovusCommandLineOption(OptionsList.Items[i]);
     if lOption.Required = true then
     begin
       Result := true;
@@ -892,7 +1060,7 @@ Var
 begin
   Result := NIL;
 
-  fNovusCommandLineOption := fOptionList.FindItem(aOptionName);
+  fNovusCommandLineOption := fOptionsList.FindItem(aOptionName);
 
   if Assigned(fNovusCommandLineOption) then
     Result := tNovusCommandLineOption(fNovusCommandLineOption);
@@ -908,7 +1076,7 @@ begin
 
   For i := 0 to fCommandList.count - 1 do
   begin
-    fCommand := fCommandList.items[i] as tNovusCommandLineCommand;
+    fCommand := fCommandList.Items[i] as tNovusCommandLineCommand;
 
     if (Uppercase(fCommand.CommandName) = Uppercase(aCommandName)) or
       (Uppercase(fCommand.ShortCommandName) = Uppercase(aCommandName)) then
@@ -931,7 +1099,7 @@ begin
   Result := fObjectList.FindItem(aClassName);
 end;
 
-function tNovusCommandLineCommand.Parse: Boolean;
+function tNovusCommandLineCommand.Parse: boolean;
 begin
   Result := Execute;
 end;
@@ -948,22 +1116,22 @@ end;
 
 constructor tNovusCommandLineCommand.Create;
 begin
-  fOptionList := tNovuslist.Create;
+  fOptionsList := tNovusList.Create;
   fErrorMessages := TStringlist.Create;
-  fObjectList := tNovuslist.Create;
+  fObjectList := tNovusList.Create;
 end;
 
 destructor tNovusCommandLineCommand.Destroy;
 begin
   fObjectList.Free;
   fErrorMessages.Free;
-  fOptionList.Free;
+  fOptionsList.Free;
   fCommandList.Free;
 end;
 
 // tNovusCommandLineOption
 function tNovusCommandLineCommand.RegisterOption(const aOptionName: string;
-  const aHelp: string; const aRequired: Boolean;
+  const aHelp: string; const aRequired: boolean;
   aCommandLineOption: tNovusCommandLineOption): tNovusCommandLineOption;
 var
   fCommandLineOption: tNovusCommandLineOption;
@@ -978,17 +1146,27 @@ begin
   fCommandLineOption.Help := aHelp;
   fCommandLineOption.Required := aRequired;
 
-  fOptionList.Add(aOptionName, fCommandLineOption);
+  fOptionsList.Add(aOptionName, fCommandLineOption);
 end;
 
-procedure tNovusCommandLineCommand.SetRequired(Value: Boolean);
+procedure tNovusCommandLineCommand.SetRequired(Value: boolean);
 begin
   fbRequired := Value;
 end;
 
-function tNovusCommandLineCommand.GetRequired: Boolean;
+function tNovusCommandLineCommand.GetRequired: boolean;
 begin
   Result := fbRequired;
+end;
+
+procedure tNovusCommandLineCommand.SetHelpCommand(Value: boolean);
+begin
+  fbHelpCommand := Value;
+end;
+
+function tNovusCommandLineCommand.GetHelpCommand: boolean;
+begin
+  Result := fbHelpCommand;
 end;
 
 procedure tNovusCommandLineCommand.SetExitCode(Value: Integer);
@@ -1001,14 +1179,19 @@ begin
   Result := fiExitCode;
 end;
 
-function tNovusCommandLineCommand.GetCommandList: tNovuslist;
+function tNovusCommandLineCommand.GetCommandList: tNovusList;
 begin
   Result := fCommandList;
 end;
 
-procedure tNovusCommandLineCommand.SetCommandList(Value: tNovuslist);
+procedure tNovusCommandLineCommand.SetCommandList(Value: tNovusList);
 begin
   fCommandList := Value;
+end;
+
+function tNovusCommandLineCommand.GetOptionsCount: Integer;
+begin
+  Result := fOptionsList.count;
 end;
 
 // tNovusCommandlineResult
@@ -1037,27 +1220,38 @@ begin
   Result := fCommands;
 end;
 
-procedure TNovusCommandLineResult.SetCommands(Value: INovusCommandLineResultCommands);
+procedure TNovusCommandLineResult.SetCommands
+  (Value: INovusCommandLineResultCommands);
 begin
   fCommands := Value;
 end;
 
-function TNovusCommandLineResult.GetIsCommandEmpty: Boolean;
+function TNovusCommandLineResult.GetIsCommandEmpty: boolean;
 begin
   Result := fbIsCommandEmpty;
 end;
 
-procedure TNovusCommandLineResult.SetIsCommandEmpty(Value: Boolean);
+procedure TNovusCommandLineResult.SetIsCommandEmpty(Value: boolean);
 begin
   fbIsCommandEmpty := Value;
 end;
 
-function TNovusCommandLineResult.GetErrors: Boolean;
+function TNovusCommandLineResult.GetIsHelpCommand: boolean;
+begin
+  Result := fbIsHelpCommand;
+end;
+
+procedure TNovusCommandLineResult.SetIsHelpCommand(Value: boolean);
+begin
+  fbIsHelpCommand := Value;
+end;
+
+function TNovusCommandLineResult.GetErrors: boolean;
 begin
   Result := fbErrors;
 end;
 
-procedure TNovusCommandLineResult.SetErrors(Value: Boolean);
+procedure TNovusCommandLineResult.SetErrors(Value: boolean);
 begin
   fbErrors := Value;
 end;
@@ -1087,6 +1281,11 @@ begin
   Result := fsHelp;
 end;
 
+procedure TNovusCommandLineResult.SetHelp(Value: String);
+begin
+  fsHelp := Value;
+end;
+
 procedure TNovusCommandLineResult.AddError(aErrorMessage: string;
   aExitCode: Integer);
 begin
@@ -1108,33 +1307,32 @@ begin
   fiExitCode := aCommand.ExitCode;
 end;
 
-
-function TNovusCommandLineResult.FindFirstCommand(aCommandName: String): INovusCommandLineResultCommand;
+function TNovusCommandLineResult.FindFirstCommand(aCommandName: String)
+  : INovusCommandLineResultCommand;
 Var
-  fNovusCommandLineResultCommands: INovusCommandLineResultCommands;
+  FNovusCommandLineResultCommands: INovusCommandLineResultCommands;
   fNovusCommandLineResultCommand: INovusCommandLineResultCommand;
 begin
   Result := NIL;
 
-  fNovusCommandLineResultCommands := Commands.GetCommands
-    ( aCommandName);
-  if Assigned(fNovusCommandLineResultCommands) then
-      Result := fNovusCommandLineResultCommands.FirstCommand;
+  FNovusCommandLineResultCommands := Commands.GetCommands(aCommandName);
+  if Assigned(FNovusCommandLineResultCommands) then
+    Result := FNovusCommandLineResultCommands.FirstCommand;
 end;
 
-function TNovusCommandLineResult.FindFirstCommandwithOption(aCommandName: String): INovusCommandLineResultOption;
+function TNovusCommandLineResult.FindFirstCommandwithOption
+  (aCommandName: String): INovusCommandLineResultOption;
 Var
-  fNovusCommandLineResultCommands: INovusCommandLineResultCommands;
   fNovusCommandLineResultCommand: INovusCommandLineResultCommand;
 begin
   Result := NIL;
 
   fNovusCommandLineResultCommand := FindFirstCommand(aCommandName);
   if Assigned(fNovusCommandLineResultCommand) then
-    begin
-      if not fNovusCommandLineResultCommand.IsCommandOnly then
-         Result := fNovusCommandLineResultCommand.Options.FirstOption;
-    end;
+  begin
+    if not fNovusCommandLineResultCommand.IsCommandOnly then
+      Result := fNovusCommandLineResultCommand.Options.FirstOption;
+  end;
 end;
 
 // tNovusCommandLineOption
@@ -1159,12 +1357,12 @@ begin
   fsOptionName := Value;
 end;
 
-procedure tNovusCommandLineOption.SetRequired(Value: Boolean);
+procedure tNovusCommandLineOption.SetRequired(Value: boolean);
 begin
   fbRequired := Value;
 end;
 
-function tNovusCommandLineOption.GetRequired: Boolean;
+function tNovusCommandLineOption.GetRequired: boolean;
 begin
   Result := fbRequired;
 end;
@@ -1179,7 +1377,7 @@ begin
   fsHelp := Value;
 end;
 
-function tNovusCommandLineOption.Execute: Boolean;
+function tNovusCommandLineOption.Execute: boolean;
 begin
   Result := true;
 end;
@@ -1224,9 +1422,9 @@ constructor TNovusCommandLineResultCommand.Create;
 begin
   fsCommandName := '';
 
-  fbIsCommand := False;
+  fbIsCommand := false;
 
-  fOptions:= TNovusCommandLineResultOptions.Create;
+  fOptions := TNovusCommandLineResultOptions.Create;
 end;
 
 destructor TNovusCommandLineResultCommand.Destroy;
@@ -1239,13 +1437,13 @@ begin
   Result := fsCommandName;
 end;
 
-
 function TNovusCommandLineResultCommand.GetOptions: INovusCommandLineResultOptions;
 begin
   Result := fOptions;
 end;
 
-procedure TNovusCommandLineResultCommand.SetOptions(Value: INovusCommandLineResultOptions);
+procedure TNovusCommandLineResultCommand.SetOptions
+  (Value: INovusCommandLineResultOptions);
 begin
   fOptions := Value;
 end;
@@ -1255,12 +1453,12 @@ begin
   fsCommandName := Value;
 end;
 
-function TNovusCommandLineResultCommand.GetIsCommandOnly: Boolean;
+function TNovusCommandLineResultCommand.GetIsCommandOnly: boolean;
 begin
   Result := fbIsCommandOnly;
 end;
 
-procedure TNovusCommandLineResultCommand.SetIsCommandOnly(Value: Boolean);
+procedure TNovusCommandLineResultCommand.SetIsCommandOnly(Value: boolean);
 begin
   fbIsCommandOnly := Value;
 end;
@@ -1270,7 +1468,8 @@ constructor TNovusCommandLineResultCommands.Create;
 begin
   fiIndex := 0;
 
-  FNovusCommandLineResultCommands:= TNovusList.Create(TNovusCommandLineResultCommand);
+  FNovusCommandLineResultCommands :=
+    tNovusList.Create(TNovusCommandLineResultCommand);
 end;
 
 destructor TNovusCommandLineResultCommands.Destroy;
@@ -1284,84 +1483,96 @@ begin
 end;
 
 procedure TNovusCommandLineResultCommands.SetItems(Value: tNovusList);
- begin
-   FNovusCommandLineResultCommands := Value;
- end;
+begin
+  FNovusCommandLineResultCommands := Value;
+end;
 
 function TNovusCommandLineResultCommands.GetIndex: Integer;
 begin
   Result := fiIndex;
 end;
 
-procedure TNovusCommandLineResultCommands.SetIndex(value: integer);
+procedure TNovusCommandLineResultCommands.SetIndex(Value: Integer);
 begin
-  fiIndex := value;
+  fiIndex := Value;
 end;
 
-procedure TNovusCommandLineResultCommands.Add(aItem: INovusCommandLineResultCommand);
+procedure TNovusCommandLineResultCommands.Add
+  (aItem: INovusCommandLineResultCommand);
 begin
   FNovusCommandLineResultCommands.Add(aItem as TNovusCommandLineResultCommand);
 end;
 
-function TNovusCommandLineResultCommands.GetCommands(aCommandName: string): INovusCommandLineResultCommands;
+function TNovusCommandLineResultCommands.GetCommands(aCommandName: string)
+  : INovusCommandLineResultCommands;
 Var
   fNovusCommandLineResultCommand: TNovusCommandLineResultCommand;
-  I: Integer;
+  i: Integer;
 begin
   Result := NIL;
 
-  if Not Exists(aCommandName) then Exit;
+  if Not Exists(aCommandName) then
+    Exit;
 
-  Result := tNovusCommandLineResultCommands.Create;
+  Result := TNovusCommandLineResultCommands.Create;
 
-  for I := 0 to Items.Count -1 do
+  for i := 0 to Items.count - 1 do
+  begin
+    fNovusCommandLineResultCommand :=
+      (Items[i] as TNovusCommandLineResultCommand);
+
+    if Uppercase(Trim(fNovusCommandLineResultCommand.CommandName))
+      = Uppercase(Trim(aCommandName)) then
+      Result.Add(fNovusCommandLineResultCommand);
+
+  end;
+
+end;
+
+function TNovusCommandLineResultCommands.FindCommandByName(aCommandName: string): INovusCommandLineResultCommand;
+Var
+  i: Integer;
+  LItem: TNovusCommandLineResultCommand;
+begin
+  Result := NIL;
+
+  for i := 0 to Items.count - 1 do
+  begin
+    LItem := TNovusCommandLineResultCommand(Self.Items[i]);
+
+    if Uppercase(Trim(LItem.CommandName)) = Uppercase(Trim(aCommandName)) then
     begin
-      fNovusCommandLineResultCommand := (Items[i] as TNovusCommandLineResultCommand);
+      Result := LItem;
 
-      if Uppercase( Trim(fNovusCommandLineResultCommand.CommandName)) = Uppercase(Trim(aCommandName)) then
-          Result.Add(fNovusCommandLineResultCommand);
-
+      break;
     end;
-
+  end;
 end;
 
 
 function TNovusCommandLineResultCommands.Exists(aCommandName: String): boolean;
-Var
-  I: integer;
-  LItem: TNovusCommandLineResultCommand;
 begin
-  Result := False;
-
-  for I := 0 to FNovusCommandLineResultCommands.Count -1 do
-    begin
-      LItem  := TNovusCommandLineResultCommand(Self.Items[i]);
-
-      if Uppercase(Trim(LItem.CommandName)) = Uppercase(Trim(aCommandName)) then
-        begin
-          Result := True;
-
-          Break;
-        end;
-    end;
+  Result := (FindCommandByName(aCommandName) <> NIL);
 end;
 
-function TNovusCommandLineResultCommands.FirstCommand: INovusCommandLineResultCommand;
+function TNovusCommandLineResultCommands.FirstCommand
+  : INovusCommandLineResultCommand;
 begin
   Result := NIL;
 
   fiIndex := 0;
 
-  if Items.Count > 0 then
-    Result := TNovusCommandLineResultCommand(Items[fiIndex ]);
+  if Items.count > 0 then
+    Result := TNovusCommandLineResultCommand(Items[fiIndex]);
 end;
 
-function TNovusCommandLineResultCommands.NextCommand: INovusCommandLineResultCommand;
+function TNovusCommandLineResultCommands.NextCommand
+  : INovusCommandLineResultCommand;
 begin
   Result := NIL;
 
-  Inc(fiIndex);
-  if fiIndex > (Items.Count -1) then
+  inc(fiIndex);
+  if fiIndex > (Items.count - 1) then
     Exit;
 
   Result := TNovusCommandLineResultCommand(Items[fiIndex]);
@@ -1389,11 +1600,11 @@ begin
   fsValue := Value;
 end;
 
-
 // TNovusCommandLineResultOptions
 constructor TNovusCommandLineResultOptions.Create;
 begin
-  FNovusCommandLineResultOptions := tNovusList.Create(tNovusCommandLineResultOption);
+  FNovusCommandLineResultOptions :=
+    tNovusList.Create(TNovusCommandLineResultOption);
 end;
 
 destructor TNovusCommandLineResultOptions.Destroy;
@@ -1411,17 +1622,39 @@ begin
   FNovusCommandLineResultOptions := Value;
 end;
 
-procedure TNovusCommandLineResultOptions.Add(aItem: INovusCommandLineResultOption);
+procedure TNovusCommandLineResultOptions.Add
+  (aItem: INovusCommandLineResultOption);
 begin
-  FNovusCommandLineResultOptions.Add(aItem as tNovusCommandLineResultOption);
+  FNovusCommandLineResultOptions.Add(aItem as TNovusCommandLineResultOption);
 end;
 
-function TNovusCommandLineResultOptions.FirstOption: INovusCommandLineResultOption;
+function TNovusCommandLineResultOptions.FirstOption
+  : INovusCommandLineResultOption;
 begin
-   Result := NIL;
+  Result := NIL;
 
-  if Items.Count > 0 then
+  if Items.count > 0 then
     Result := TNovusCommandLineResultOption(Items[0]);
+end;
+
+function TNovusCommandLineResultOptions.FindOptionByName(aOptionName: String): INovusCommandLineResultOption;
+Var
+  i: Integer;
+  LItem: TNovusCommandLineResultOption;
+begin
+  Result := NIL;
+
+  for i := 0 to Items.count - 1 do
+  begin
+    LItem := TNovusCommandLineResultOption(Self.Items[i]);
+
+    if Uppercase(Trim(LItem.OptionName)) = Uppercase(Trim(aOptionName)) then
+    begin
+      Result := LItem;
+
+      break;
+    end;
+  end
 end;
 
 end.